# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Aldb Example API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api",
    "paths": {
...
        "/alga/anno": {
            "get": {
                "description": "藻类图像获取标注",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aldb"
                ],
                "summary": "GetAnnotationByAlga",
                "parameters": [
                    {
                        "type": "string",
                        "description": "藻类名称",
                        "name": "alga",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code, msg, data}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/user/anno": {
            "get": {
                "description": "用户获取个人标注",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aldb"
                ],
                "summary": "GetAnnotationByUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户邮箱",
                        "name": "user",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code, msg, data}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/user/delete": {
            "get": {
                "description": "删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aldb"
                ],
                "summary": "DeleteUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code, msg, data}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aldb"
                ],
                "summary": "GetUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{code, msg, data}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    }
}